feat: combine all the trees into the package "trees" and clean code
