+++ app/src/main/kotlin/dataBase/Json.kt
@@ -1,6 +1,6 @@
 package dataBase
 
-import BinTree
+import trees.BinTree
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import com.fasterxml.jackson.module.kotlin.readValue
 import java.io.File
