+++ lib/src/main/kotlin/trees/AVLTree.kt
@@ -1,3 +1,5 @@
+package trees
+
 open class AVLTree<Key : Comparable<Key>, Value> : BalanceTree<Key, Value> {
     protected class AVLNode<Key : Comparable<Key>, Value>(
         key: Key,
