+++ lib/src/main/kotlin/trees/BalanceTree.kt
@@ -1,3 +1,5 @@
+package trees
+
 abstract class BalanceTree<Key : Comparable<Key>, Value> : BinTree<Key, Value> {
     constructor() : super()
     constructor(key: Key, value: Value) : super(key, value)
