+++ lib/src/main/kotlin/trees/BinTree.kt
@@ -1,3 +1,5 @@
+package trees
+
 import java.util.LinkedList
 import java.util.Queue
 import kotlin.math.abs
