+++ lib/src/main/kotlin/trees/RBTree.kt
@@ -1,3 +1,5 @@
+package trees
+
 open class RBTree<Key : Comparable<Key>, Value> : BalanceTree<Key, Value> {
     companion object {
         const val RED = false
