+++ lib/src/test/kotlin/trees/BSTreeTest.kt
@@ -1,3 +1,5 @@
+package trees
+
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Nested
